hermetic and integration mark
